Save data to database using /api/match/espn/getPastResults
