+++ routes/match.js
@@ -3,5 +3,6 @@ const router = express.Router();
 const matchController = require("../controllers/match");
 
 router.post("", matchController.createMatch);
+router.get('/espn/getPastResults', matchController.getPastResults);
 
 module.exports = router;
